Create a base class for PortalNavigationName in case we want others in the future.
