+++ Lombiq.HelpfulLibraries/Docs/Navigation.md
@@ -1,5 +1,14 @@
 # Navigation Documentation
 
+
+
+## NavigationProviderBase
+
+An abstract base class for creating reducing boilerplate in `INavigationProvider` for checking the name parameter and if the user is authenticated. 
+
+
 ## PortalNavigationProviderBase
 
-An abstract base class for creating home page menu structure using the `portal` navigation name. If you use the _Lombiq.BaseTheme_ it automatically displays the generated menu as a widget in the _Navigation_ zone.
+An abstract base class derived from `NavigationProviderBase` for creating home page menu structure using the `portal` navigation name. If you use the _Lombiq.BaseTheme_ it automatically displays the generated menu as a widget in the _Navigation_ zone.
+
+
